% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age5_to_age1.R
\name{age5_to_age1}
\alias{age5_to_age1}
\title{Konversi jumlah penduduk umur 5 tahunan ke 1 tahunan}
\usage{
age5_to_age1(oag, value)
}
\arguments{
\item{oag}{suatu numerik/angka untuk kelompok umur terbuka, misal 80 adalah untuk 80+}

\item{value}{vektor numerik untuk jumlah penduduk 5 tahunan}
}
\description{
Fungsi untuk memecah jumlah penduduk umur 5 tahunan menjadi 1 tahunan.
}
\details{
Pemecahan jumlah penduduk umur 5 tahunan menjadi 1 tahunan dengan menggunakan koefisien sprague.
}
\examples{
# library(dplyr)
# library(magrittr)
# library(estPUSek)
#
# data(pop_age1_Tot)
#
# menghitung/agregasi jumlah penduduk umur5 5 tahunan
# df <- pop_age1_Tot \%>\%
#   filter(kode_provinsi == 0, tahun == 2020) #ambil untuk Indonesia, tahun 2020
# pop_age5 <- df \%>\%
#   mutate(umur5 = cut(umur,c(seq(0,100, by=5)), include.lowest = TRUE,
#                     right = FALSE, labels = FALSE),
#          umur5 = ifelse(umur5 > 17, 17, umur5),
#          umur5 = factor(umur5, labels=seq(0,80,5))
#   ) \%>\%
#   group_by(umur5) \%>\%
#   summarise(Jumlah = sum(values)) \%>\%
#   ungroup()
#
# estimasi jumlah penduduk umur 1 tahunan
# pop_age1_est <- age5_to_age1(oag=80,pop_age5$Jumlah)
#
# sandingan
# cbind(df$values, pop_age1_est)

}
\references{
Siegel Jacob S, Swanson David A (eds.) (2004). The Methods and Materials of Demography, 2 edition. Elsevier Academic Press, California, USA, San Diego, USA
Sprague, T. B. (1880). Explanation ov a New Formula for Interpolation. Journal of the Institute of Actuaries and Assurance Magazine, 22(4), 270–285.
Riffe T, Aburto J, Kashnitsky I, Alexander M, Pascariu M, Hertog S, Fennell S (2023). \emph{DemoTools: Standardize, Evaluate, and Adjust Demographic Data}. R package version 01.13.80.
}
