% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPUSek.R
\name{run.estPUSek}
\alias{run.estPUSek}
\title{Menjalankan aplikasi estPUSek}
\usage{
run.estPUSek(host = NULL, ...)
}
\arguments{
\item{host}{argumen untuk \code{\link[shiny:runApp]{runApp}}. Default "0.0.0.0".}

\item{...}{argumen tambahan untuk \code{\link[shiny:runApp]{runApp}}.}
}
\description{
Menjalankan estPUSek, aplikasi berbasis shiny untuk estimasi jumlah penduduk usia sekolah dan umur 1 tahunan.
}
\details{
Deskripsi fitur-fitur tersedia di halaman bantuan aplikasi.
}
\examples{
run.estPUSek()

}
\author{
Ari Purwanto Sarwo Prasojo, mengadaptasi paket wppExplorer.
}
